refactor: rename useCookieConsent and streamline cookie consent logic
