+++ .gitignore
@@ -14,6 +14,7 @@ i18n.cache
 .claude
 .cursor
 .codex
+.vscode
 
 # Ralph config (may contain API key)
 .cursor/ralph-config.json
